fix: setting menu-surface position
The issue is that setting style property to "undefined" doesn't reset
the current value. Because of that, depending on the current position of
the menu, it is positioned using both "top" and "bottom" properties. The
result is that menu changes its size instead of just being positioned.

Fixed by setting to empty string.